requirements.txt
streamlit~=1.37.1
pandas~=2.2.2
numpy~=1.26.4
matplotlib~=3.9.2
seaborn~=0.13.2
torch
torchvision~=0.18.1
transformers~=4.48.1
shap~=0.46.0
pillow~=10.4.0